+++ src/extensions/proxy/proxyconn.cpp
@@ -483,7 +483,7 @@ int ProxyConn::readvSsl(const struct iovec *vector,
 }
 
 
-int ProxyConn::readv(struct iovec *vector, size_t count)
+int ProxyConn::readv(struct iovec *vector, int count)
 {
     int len;
     int total = 0;
@@ -938,7 +938,7 @@ void ProxyConn::dump()
             "Response header: %d, total: %d bytes received in %ld seconds,"
             "Total processing time: %ld.",
             getState(), getEvents(), m_iReqHeaderSize,
-            (long long)m_iReqBodySize, (long long)m_iReqTotalSent, 
+            (long long)m_iReqBodySize, (long long)m_iReqTotalSent,
             m_iRespHeaderRecv, m_iRespRecv,
             (m_lReqSentTime) ? time(NULL) - m_lReqSentTime : 0,
             time(NULL) - m_lReqBeginTime);
